+++ trunk/src/it/greentone/gui/dialog/EditOperationTypeTypologyDialog.java
@@ -14,6 +14,7 @@ import javax.swing.JButton;
 import javax.swing.JDialog;
 import javax.swing.JLabel;
 import javax.swing.JList;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTextField;
@@ -157,11 +158,17 @@ public class EditOperationTypeTypologyDialog extends JDialog
 						  getOperationTypeTypologyJList().getSelectedIndex();
 						if(selectedIndex > -1)
 						{
-							OperationTypeTypology operationTypeTypology =
-							  operationTypeTypologyService.getAllOperationTypeTypologies()
-							    .get(selectedIndex);
-							operationTypeTypologyService
-							  .deleteOperationTypeTypology(operationTypeTypology);
+							int confirmDialog =
+							  JOptionPane.showConfirmDialog(getContentPane(), resourceMap
+							    .getString("editOperationTypeTypology.Action.confirmMessage"));
+							if(confirmDialog == JOptionPane.OK_OPTION)
+							{
+								OperationTypeTypology operationTypeTypology =
+								  operationTypeTypologyService.getAllOperationTypeTypologies()
+								    .get(selectedIndex);
+								operationTypeTypologyService
+								  .deleteOperationTypeTypology(operationTypeTypology);
+							}
 						}
 					}
 				});
